// kitchen-planner-fe-main/src/app/api/api-kms-planner-masterdata/model/userRegistration.ts
/**
 * KMS - Kitchen Management System API
 *
 * Contact: anas-ur-rasheed.abid@tu-ilmenau.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserRegistration { 
    firstName: string;
    lastName: string;
    email: string;
    password: string;
    role: UserRegistration.RoleEnum;
}
export namespace UserRegistration {
    export const RoleEnum = {
        User: 'user',
        Chef: 'chef',
        Admin: 'admin'
    } as const;
    export type RoleEnum = typeof RoleEnum[keyof typeof RoleEnum];
}